% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_cpp}
\alias{add_cpp}
\title{Add Two Numbers}
\usage{
add_cpp(x, y)
}
\arguments{
\item{x}{numeric; the first number}

\item{y}{numeric; the second number}
}
\description{
Return the sum of two numbers.
}
