% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumC}
\alias{sumC}
\title{Adds a vector of numbers}
\usage{
sumC(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
value The sum of the numbers
}
\description{
Return the sum of all numbers in a numeric vector.
}
