% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_cpp}
\alias{add_cpp}
\title{Add Two Numbers}
\usage{
add_cpp(x, y)
}
\arguments{
\item{x}{is the first numeric value}

\item{y}{is the second numeric value}
}
\description{
Return the sum of two numbers.
}
