% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_cpp}
\alias{add_cpp}
\title{Add Two Numbers}
\usage{
add_cpp(x, y)
}
\arguments{
\item{x}{A number as a double}

\item{y}{A number as a double}
}
\value{
value The sum of the two inputs
}
\description{
Return the sum of two numbers.
}
